move reduce_events() to new module reduce.py
