+++ tensorboard_reducer/__init__.py
@@ -1,6 +1,7 @@
 from tensorboard_reducer.load import load_tb_events as load_tb_events
 from tensorboard_reducer.main import main as main
 from tensorboard_reducer.main import reduce_events as reduce_events
+from tensorboard_reducer.reduce import reduce_events as reduce_events
 from tensorboard_reducer.write import write_data_file as write_data_file
 from tensorboard_reducer.write import write_df as write_df
 from tensorboard_reducer.write import write_tb_events as write_tb_events
