+++ tensorboard_reducer/main.py
@@ -2,49 +2,10 @@ from __future__ import annotations
 
 from argparse import ArgumentParser
 from importlib.metadata import version
-from typing import Sequence
-
-import pandas as pd
 
 from tensorboard_reducer.load import load_tb_events
+from tensorboard_reducer.reduce import reduce_events
 from tensorboard_reducer.write import write_data_file, write_tb_events
-
-
-def reduce_events(
-    events_dict: dict[str, pd.DataFrame],
-    reduce_ops: Sequence[str],
-    verbose: bool = False,
-) -> dict[str, dict[str, pd.DataFrame]]:
-    """Perform numpy reduce operations along the last dimension of each array in a
-    dictionary of scalar TensorBoard event data. Each array (1 per run) enters this
-    function with shape (n_steps, n_runs) and it returns a dict of len(reduce_ops)
-    subdicts each with keys named after scalar quantities (loss, accuracy, etc.) holding
-    arrays with shape (n_steps,).
-
-    Args:
-        events_dict (dict[str, pd.DataFrame]): Dict of arrays to reduce.
-        reduce_ops (list[str]): Names of numpy reduce ops. E.g. mean, std, min, max, ...
-        verbose (bool, optional): Whether to print progress. Defaults to False.
-
-    Returns:
-        dict[str, dict[str, pd.DataFrame]]: Dict of dicts where each subdict holds one
-            reduced array for each of the specified reduce ops, e.g.
-            {"loss": {"mean": arr.mean(-1), "std": arr.std(-1)}}.
-    """
-    reductions: dict[str, dict[str, pd.DataFrame]] = {}
-
-    for op in reduce_ops:
-        reductions[op] = {}
-
-        for tag, df in events_dict.items():
-            reductions[op][tag] = getattr(df, op)(axis=1)
-
-    if verbose:
-        print(
-            f"Reduced {len(events_dict)} scalars with {len(reduce_ops)} operations:"
-            f" ({', '.join(reduce_ops)})"
-        )
-    return reductions
 
 
 def main(argv: list[str] | None = None) -> int:
